+++ gems/markdown-merge/spec/markdown/merge/code_fence_dialect_spec.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+require_relative '../../spec_helper'
+
+RSpec.describe Markdown::Merge do
+  it 'routes JSON5 fenced blocks to the JSON merge family and dialect' do
+    expect(described_class.code_fence_family('json5')).to eq('json')
+    expect(described_class.code_fence_dialect('json5', 'json')).to eq('json5')
+  end
+end